requirements/common.txt
django==1.6.5

pillow==2.5.1
django-imagekit==3.2.1

django-guardian==1.2.2

django-taggit==0.12

django-crispy-forms==1.4.0

django-cleanup==0.1.10

django-allauth==0.17.0

django-kronos==0.4

cssselect==0.9.1
lxml==3.3.5
pysolr==3.2.0
django-haystack==2.2.0

django-ratelimit==0.4.0

redis==2.10.2

sorl-thumbnail==11.12.1b

dicttoxml==1.5.5

django-recaptcha==1.0

geonode-agon-ratings==0.3.1

dialogos==0.4
